requirements-dev.txt
-r requirements.txt
Flask-Test==0.1.5
pytest==2.4.2
flexmock==0.9.7
blinker==1.3
Flask-Login==0.2.7
factory_boy==2.2.1
